Initial version
This is a port of the library pieces of ogud's DNSSEC-resolver-check
code from java to python.

Run UI_DRC.py, and it will detect your resolvers and report on them. Or
add more resolvers (like, run UI_DRC.py 8.8.8.8, and it will report on
the resolver there - in this case, google's open resolver, which is a
validator.

There is more documentation at
https://github.com/ogud/DNSSEC-resolver-check/